Create modify_c_lattice.py
This script reformats both the lattice vectors and atomic coordinate lines using fixed‐width formatting (15 characters wide with 10 decimal places) so that the decimal points line up neatly. In addition, the script modifies the c lattice vector (line 5) by setting its x and y components to zero and adding a user‐specified value (delta_z) to its z component. It also checks that the coordinate system indicator (on line 7 or 8) begins with “c” or “k” (for Cartesian) and inserts a chemical symbols line if one is missing.